Bugfix: Rename CR kind to KnativeEventing (#48)
* Rename of Eventing to KnativeEventing

* Generating code for KnativeEventing

* README.md remove example with short name ke

Plus a lint fix for README.md

* Remove short name for Knative Eventing

There's a problem, that if we have older version of knative
eventing operator installed, this short name is occupied by
the CRD eventings.operator.knative.dev. After we applying
the new CRD knativeeventings.operator.knative.dev also with
this same short name, the terrible thing will happen: none
of the old or the new CR can be found by the operator
controller.

Typically you won't deal with those resources on a daily
basis, occupying a 2-letter slot seems to be intrusive
and not really justified.

Taking above to account, removing that short name seams
most reasonable.