Use `escalate` and aggregation to minimize RBAC.
By using the `escalate` verb on clusterroles, we allow the Operator
to have the ability to create any clusterrole. This prevents us from
needing to grant it `cluster-admin`.

Additionally, by having the Operator create an aggregated clusterrole
for itself (aggregating all clusteroles created by Knative Eventing),
we pick up any permissions that the upstream dictate as necessary
by any component.

Any roles that are not created by upstream, but which are bound as
part of install are added with the `bind` verb.

Overall, this should keep the Operator roughly operating only on
Knative Eventing resources (or things which can be operated
on by upstream).