Delete (Cluster)Role(Bindings) as a final cleanup step. (#141)
* Delete (Cluster)Role(Bindings) as a final cleanup step.

This allows us to take advantage of the permissions granted to cluster admins
when performing cleanup. If the approach in #109
is approved, this will become necessary. Additionally, it makes sense to remove roles as the final
step to allow human Operators to modify any resources they may have permissions on as a result of the Knative installation (that is, we should not remove any access until we are 'almost done' cleaning up).

* Remove redundant All(), fix conditional.