Restrict RBAC for Operator (#109)
* Use `escalate` and aggregation to minimize RBAC.

By using the `escalate` verb on clusterroles, we allow the Operator
to have the ability to create any clusterrole. This prevents us from
needing to grant it `cluster-admin`.

Additionally, by having the Operator create an aggregated clusterrole
for itself (aggregating all clusteroles created by Knative Eventing),
we pick up any permissions that the upstream dictate as necessary
by any component.

Any roles that are not created by upstream, but which are bound as
part of install are added with the `bind` verb.

Overall, this should keep the Operator roughly operating only on
Knative Eventing resources (or things which can be operated
on by upstream).

* Update to correct namespaces for Operator SA and binding.

* Readd daemon/replica/stateful/sets.

* Removing 'specific to this operator' and blanket role.

* Install from the manifest in a strict order.

This order, Roles -> RoleBindings -> The Rest, prevents the operand from
needing to structure their manifest such that this ordering exists, while
allowing the Operator to use the bootstrapping mechanism described in
https://github.com/knative/eventing-operator/pull/109 to 'escalate' itself
into management of the Knative Eventing installation without a *** clusterrole.

* Fixing golint

* Role in eventing namespace is no longer required due to role aggregation.

* TODO suggested improvement.

* Final update for pretty diff between role in serving and eventing.

* Final final update. Add copyright header.

* Another copyright, serving -> eventing.

And knative-eventing namespace to knativeeventings namespaces.